backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0
